flat verifier key

+++ packages/loopring_v3/contracts/iface/IExchangeV3.sol
@@ -34,7 +34,7 @@ abstract contract IExchangeV3 is IExchange
     );
 
     event AccountCapacityIncreased(
-        uint accountTreeDepth
+        uint8 accountTreeDepth
     );
 
     event Shutdown(
